code/ocr/requirements-stub.txt
flask
pillow
requests
numpy